Soccer dataset cleaning
The cleaning work done on the soccer dataset involved several steps to prepare the data for analysis.

The height and weight columns were converted from imperial units to metric units. This was done using custom functions that split the height and weight strings into feet/inches and pounds, respectively, and then converted them to centimeters and kilograms.

New column with a contract duration has been created based on a contract start and end dates. Contract status has been added for "free" and "on loan" cases.

Finally, the hits column was also converted from a string to a float, using a custom function that checked for the presence of "K" in the string and then multiplied the value by 1000.

Overall, these cleaning steps helped to ensure that the soccer dataset was in a consistent and usable format for analysis.